+++ src/EventSourcing/Logitar.EventSourcing.EntityFrameworkCore.SqlServer/Migrations/20230804160024_CreateEventTable.cs
@@ -18,7 +18,7 @@ namespace Logitar.EventSourcing.EntityFrameworkCore.SqlServer.Migrations
                     EventId = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    ActorId = table.Column<string>(type: "nvarchar(255)", maxLength: 255, nullable: true),
+                    ActorId = table.Column<string>(type: "nvarchar(255)", maxLength: 255, nullable: false),
                     OccurredOn = table.Column<DateTime>(type: "datetime2", nullable: false),
                     Version = table.Column<long>(type: "bigint", nullable: false),
                     DeleteAction = table.Column<string>(type: "nvarchar(255)", maxLength: 255, nullable: false),