+++ src/EventSourcing/Logitar.EventSourcing.Relational/EventEntity.cs
@@ -19,7 +19,7 @@ public class EventEntity : IEventEntity
   /// <summary>
   /// Gets or sets the identifier of the actor who triggered the event.
   /// </summary>
-  public string? ActorId { get; set; }
+  public string ActorId { get; set; } = string.Empty;
   /// <summary>
   /// Gets or sets the date and time when the event occurred.
   /// </summary>