+++ src/EventSourcing/Logitar.EventSourcing.SqlServer/Scripts/20230804160024_CreateEventTable.sql
@@ -14,7 +14,7 @@ GO
 CREATE TABLE [Events] (
     [EventId] bigint NOT NULL IDENTITY,
     [Id] uniqueidentifier NOT NULL,
-    [ActorId] nvarchar(255) NULL,
+    [ActorId] nvarchar(255) NOT NULL,
     [OccurredOn] datetime2 NOT NULL,
     [Version] bigint NOT NULL,
     [DeleteAction] nvarchar(255) NOT NULL,
@@ -48,7 +48,7 @@ CREATE INDEX [IX_Events_Version] ON [Events] ([Version]);
 GO
 
 INSERT INTO [__EFMigrationsHistory] ([MigrationId], [ProductVersion])
-VALUES (N'20230710193739_CreateEventTable', N'7.0.8');
+VALUES (N'20230804160024_CreateEventTable', N'7.0.9');
 GO
 
 COMMIT;