+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implemented query JOINs.
 - Added AggregateRoot metadata.
+- Created a struct for actor identifiers.
 
 ### Changed